pyproject.toml
[tool.poetry]
name = "chatbot-llama-index"
version = "0.1.0"
description = ""
authors = ["Alisson Pereira Anjos <Alisson.Anjos@valgenesis.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.10.12"
openai = "^1.3.5"
llama-index = "^0.9.6.post2"
pypdf = "^3.17.1"
python-decouple = "^3.8"
chromadb = "^0.4.18"
weaviate-client = "^3.25.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
